Remove globals.inc, reserve R0-R15 for scratch space in zero page
I don't want to lean on the idea of globals.inc as a shortcut, as
it seems like it will encourage poor organization strategies
long term. Instead I will be more explicit with which symbols I
import and export. If one file has a lot of exports (like zero
page does), it can have a matching .inc with those imports to
ease its use.