The converter is pre-multiplying the data, so don't do it here
This saves 4 cycles for each iteration of the loop. Right now
I can't think of any way to improve these loops without doing an
unroll of the data, but I'm sure there is a way to squeeze a few
more cycles out of them that I haven't thought of. Optimization
later, working first.