Reload IRQ counter during scanline 0 to avoid flicker on even/odd frames
Turns out MMC3's IRQ counter is touchy about the skipped idle cycle
(or something) that occurs at the end of the prerender scanline. When
backgrounds are fetched from $1000, this causes every other frame to
miscount exactly one scanline, leading to a jittering glitch. We work
around this by using a sprite zero hit, ending NMI with a spinwait to
signal the end of vblank, then a bit more timed code to land shortly
after the start of scanline 0. At this point the problematic skipped
cycle is behind us, so we force an IRQ counter reload, and from this
point on the scanline counter works correctly for the remainder of
the frame.

This has resolved the flicker glitch on real hardware and Mesen, but
the parameters need to be tuned so that the scroll split is smooth
when switching modes.