Use NMI spinwait instead of MMC3 IRQ for short mid-screen splits
The MMC3 IRQ is unreliable for the first 3-4 scanlines for whatever
reason. In this case we need to manually configure the playfield
split using a bit of timed code, to handle the top 3 scroll
positions manually. These are some seriously magic numbers, but
they land the required PPUADDR updates in the appropriate location
without any visible glitches.