Add an explicit squish pose and use that instead of a y offset
Turns out it's a bad idea to have 8x8 hardware tiles vertically
overlap, as this effectively doubles the "cost" for this sprite
in terms of a single scanline in the overlap region. We should
generally avoid this as an art constraint, as it otherwise leads
to ugly horizontal artifacts when even a few sprites are in the
same horizontal line together.