Implement a basic "follow entity" camera, with a bit of lerp
This is pretty close to the framework I want to use for the final
camera, but it could be tweaked and improved quite a bit yet.
It works well for now, and does finally refuse to allow the map
to be scrolled past its extreme edges. No more glitchy sprites
in those cases.