Explicitly inc/dec scroll registers, rather than blind copy
This is a blind attempt to fix an uncommon bug with the scrolling
logic that is occasionally causing the row / column trackers on
each edge to desync. I haven't been able to consistently reproduce
this bug, but the theory is that somehow the blind copy was getting
confused on frames when a horizontal *and* vertical shift were
queued up simultaneously.

This code is really how we should be operating *anyway*, since it
nudges the "current" registers in the direction of what actually
happened with the PPU copies, and disallows unusually large
jumps of 2 or more if the camera code (incorrectly) tries to
scroll too fast.