Add an orange-tinted version of the skull tileset
The converter seems to be working properly here, not duplicating
the CHR tiles, and picking up the attribute index. That's a
bonus, less of a bonus is that doing this for large sets of tiles
is *incredibly* tedious. A custom editor down the line will help
tremendously with this; Tiled is showing its limits already.