Disable spinwait on sprite zero flag, to test a broken MMC3 IRQ
We're not sure exactly what's going wrong, but if we configure
MMC3 IRQ after the prerender scanline, both Mesen and PowerPak
MMC3 implementations produce this strange jitter, as though the
scanline count is off by one on every other frame. This branch
is intended to isolate this behavior, and facilitate testing on
real hardware to confirm whether it is a bug on the real thing,
or a quick of faulty emulation.