Get the new map data loading through the compression setup
This looks glitched because the tileset is still in premultiplied
format, but the tile data is expecting normal indexing. Next we'll
need to rework the tileset exporter, then adjust the scrolling
routine so that it can work with 4 parallel lists of tileset data
instead of the one big block it's been using so far.