Encode the CHR bank as part of the tileset, using ca65 linker features
Basically we rely on symbol import/export and unique filenames
for tilesets, and use this to have the assembler work out where
in memory a given CHR pattern ends up. The segment config section
details the appropriate bank number, so we can have the map loading
function load this and stash it somewhere, and then apply it during
render. This *completely* automates the process, allowing it to
simply be part of the map. Highly convenient :)