Relocate boxgirl's sprite, so that y=0 is the ground level
I think this is the pattern I want to use for all entity sprites,
at least for those that are going to interact with the 3D terrain.
This lets me specify a full hitbox for the ground region using a
single byte and minimal modification to apply that byte. Ground
line segments (the usual case) will be specified as just 3 bytes
total.