Rework the map to dodge most problem spots with the bgPriority system
The top of the map is still questionable, but things are mostly
working. Hidden surfaces require a *lot* of headroom to deal with
jump height and all that.

I think some of that jump height clearance could be reduced, but we'll
have to see. There are definitely some level design gnargles that are
simply unavoidable with this setup.