Rework the test 3D map to have fewer visual glitches
We need a *ton* of vertical headroom to accomodate jumping, and
this still isn't quite enough to do the effect justice. I'm nearly
certain I want to try a different approach at this point, but
I want to get a map with bridges made before I settle on a strategy.