Fully implement LZ77-based compression for maps and tilesets
Yay! The scaffolding here should make it very easy to support other
compression schemes down the line. In particular I'd like to try
a lz77 variant with 8bit pointers and 8bit offsets, to see if that
helps. I imagine allowing maps to reference more than a single
row's worth of data in the sliding window is worth a try, and if
nothing else that format should support quite long data runs for
RLE.

Anyway, super happy with this. It isn't as slow to decode as I
expected, maybe 5-10 frames. I need to get map porting working so
this can be measured and tested in a game context, but it "feels"
fast enough.