Fix a silly bug with height adjustments
We now only apply a height adjustment if both points are valid. If
either point results in a push, we ignore height adjustments for that
frame. This works because collision is performed per axis, and any
push, at normal game speeds, causes the player to have effectively
not changed tiles at all on that axis. Since their previous height
is valid, we keep it unconditionally. Fixes a zip bug when colliding
against height-differentiated corners in the north and south directions.