Create a new map to test a new feature: bridges!
This is mostly a visual feature, as the engine has been designed
to handle these from the beginning. Horizontal bridges work great,
vertical ones... not so much. These are going to be a major
problem if we ever need them. We will almost certainly need to use
sprites to cheat the layer priority on the top end.