Implement a "soft NMI disable", to continue running the sound engine
Note: this does not currently play nice with vram_buffer emptying,
which really needs to not be interrupted to avoid crashes, due to
stack manipulation. As a result, the plan is to rework map init
so that it does not use the popslide technique when drawing the
initial viewport. Until we do that, a slight tempo pause during
level loading is unavoidable.