Replace the basic HUD with something more generally useful
This tileset comes equipped with a full font, not that we're using
it yet. We should be able to use this to hack together display for
energy and health and stuff while we are still knocking about
with the basic mechanics, before we commit to spending the time to
make real art.