Try a vertical jumping setup with 6-high platforms and shadows
This strictly works in a technical sense, but it's at the limit
of what feels comfortable in terms of gameplay. Notably the
ground shadow is offscreen to the bottom, and while we could
compensate for that by moving the camera "down" slightly, it's
a hack and it'd be better to avoid this setup in the first
place.