Implement a hitstun state for the engine
This basically freeze frames the action, but not the camera, for
X number of frames. Useful to give more impact to hits, generally.

Works by pausing the entire engine, so use quite sparingly. Ideally
only for player-invoked actions, like special abilities or collisions
with enemies.

For added oomph, combine with a small amount of screen shake.