Correct an issue with corrupted sprite state on hardware
Turns out if we disable rendering for the palette swap, we
corrupt the internal pointer used to manipulate the PPU's copy of
the 8 sprites for the next scanline. "Who cares, we aren't going
to render those this frame?" Ah yes, but if we *leave* that pointer
corrupted, then the next OAM DMA fails. Because PPU things.

This turns out to only affect our dialog transition state, which was
disabling rendering immediately after the palette switch, and thus
keeping the corruption. As long as we render something (even if
it is just backgrounds) after doing the palette swap, the corruption
clears up and the next OAM DMA succeeds without issue, in Mesen
while emulating all glitches, and on real hardware.

This does mean we'll need to be more careful with our HUD palettes,
and not actually rely on the blank CHR bank for hiding rendering
sluggishness.