To test an MMC3 corruption behavior, never render the dialog
This disables rendering at a poor time (early hblank) during the
palette swap, then *leaves* it disabled, which reliably trips
OAM corruption. The corruption on following scanlines is reasonably
well understood, but the post-OAM DMA corruption is not.

The community working theory is that this corruption should be
causing row duplication, with a random row in OAM being overwritten
with the contents of row 0. However on real hardware, the effect
of this corruption *also* appears to cause MMC3 to miscount one
scanline, which could only happen if a pattern is fetched from $0000
in 8x16 mode. There is no working explanation for this incorrect
fetch, which is the stated purpose of this test. :)