Pull in the Y distortion generator, hook up an underwater effect
This *almost* works! But it's having some trouble around the scroll
seam and I think that's because any split which would *cross* that
seam would display some of the HUD, which the prototype wasn't ever
set up to handle. (It assumed scroll splits would wrap around to
the top). I think it's an easy fix in terms of logic, but it may
be slow; we'll have to play with it.