Define MAX_METASPRITES, be a bit safer with loops
Old me decided that relying on bcs was a great way to check if
we were done with the loop. Unfortunately that results in running
right off the end of the table if it's smaller than 256 bytes,
which it is, and the only reason that didn't crash is that we
haven't had a level try to spawn more than max sprites yet.