When teleporting to a map with a music variant, queue that up early
The variants all have a delay, since they need to wait for the music
to get to an appropriate safe spot to switch gears. Since there is
always some sort of map transition, we can trigger the variant a tiny
bit early, and increase the chances of the variant being done by the
time the transition animation (and any accompanying SFX) are finished.
This helps to hide the seams and make the whole experience feel more
polished.