Rework dash speed, duration, and particles while underwater
This looks a lot more correct. It's slightly longer range, but
that doesn't bug me; the player is very likely to have upgraded
either the dash or the jump by the time they can go underwater,
so we'll be past the point of short gaps acting as a gating
mechanic.