Cleanup the underwater test map, add a more purple-y palette
Combined with blue emphasis this looks much better. In particular,
avoid large regions of bright (low row) palette seems to be ideal,
we can reserve those hues for entities and particles instead.

Also noteworthy, we want to avoid large regions of solid color if
at all possible, because that makes all the work that we're doing on
that hblank distortion essentially vanish.