Mooostly get ramp collision rejection in place, except there's bugs
The major bug is more of a behavioral side effect I want: collision
should only ever reject a move that would change the tile coordinate.
If we remain within the tile, we want to *allow* that move, so that
ramp pushup can apply. This is what makes ramps work at all, but
it also happens to be a potential optimization, so let's work on that.

We expanded ZP scratch to 32 bytes to have some more variables. I'm
not convinced all of these are strictly necessary, but I definitely
wanted some more breathing room for complex logic.