Rework dashing to perform several physics steps per frame
This resovles some issues trying to dash up uneven surfaces, which
was previously causing the player to "catch" on the edge of the
surface depending on their alignment when starting the dash. That
felt bad, and I suspect lag during dash frames (while the player
is functionally invincible) will feel much less bad, so... poison
picked and all that.