Fix a race condition on lag frames when performing far calls
I thought of the MMC3 register problem and had that solved, but I
failed to consider the state of CurrentBank, and ended up writing
to it in a bad order. An NMI in exactly the wrong place could
restore a stale CurrentBank when it was done, causing a crash.
Oops!