Hookup event flag management to overlay processing, cleanup logic
Right now this results in one-time switch pressing, which is then
preserved but not *applied* when loading the level. The next time
you enter the room, if you didn't un-set the trap yourself, the
state gets broken and you can no longer press the switch. (oops!)

Next task is of course to write the logic that applies all relevant
overlays on a map when loading into it for the first time, right
before the initial render.